Last commit before attempting 2 major changes.
These changes are: make the pipeline BIDS compatible, and attempting to get tensor images to properly processed. The changes since the last commit are, as per my MO, uncertain.  However it should be noted that the Hex (Sills) study and the majority of the Heritabiliy study were processed before this date.